feat(geyser): 统一喷口目录与参数链路
- 新增统一喷口目录、分类和参数提供器
- 改为按模板实体展开喷口坐标并透出参数来源与能力元数据
- 更新 sidecar 协议与预览前端以消费新的喷口详情结构

+++ src/Batch/SidecarProtocol.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <string>
 #include <vector>
 
@@ -74,7 +75,7 @@ struct SidecarPreviewRequest {
     std::string jobId;
     int worldType = 0;
     int seed = 0;
-    int mixing = 0;
+    uint64_t mixing = 0;
     PreviewTarget target = PreviewTarget::Primary;
 };
 
@@ -82,7 +83,7 @@ struct SidecarPreviewGeyserDetailsRequest {
     std::string jobId;
     int worldType = 0;
     int seed = 0;
-    int mixing = 0;
+    uint64_t mixing = 0;
     int worldHeight = 0;
     std::vector<GeyserSummary> geysers;
     PreviewTarget target = PreviewTarget::Primary;
@@ -97,7 +98,7 @@ struct SidecarWorldReportRequest {
     std::string jobId;
     int worldType = 0;
     int seed = 0;
-    int mixing = 0;
+    uint64_t mixing = 0;
 };
 
 struct SidecarCancelRequest {
